Change rating score to be a float

+++ core/database/src/main/java/com/mmorikawa/book_recommender/core/database/model/RatingEntity.kt
@@ -21,7 +21,7 @@ data class RatingEntity(
     @ColumnInfo(name = "book_id")
     @PrimaryKey
     val bookId: Int,
-    val score: Int,
+    val score: Float,
     @ColumnInfo(name = "rating_text")
     val ratingText: String,
     @ColumnInfo(name = "date_created", defaultValue = "CURRENT_TIMESTAMP")
